feat[saveGame]: success load endgame screen based on previous game data

+++ lib/src/ui/game_screen/controller/setup_save_game.dart
@@ -4,8 +4,7 @@ import '../../../data_source/local_db/key_value/modal/save_game_model.dart';
 import '../../../utils/constants.dart';
 import '../../../utils/logger.dart';
 
-class SetupSaveGame{
-
+class SetupSaveGame {
   final _logger = Logger()
       .setDebugEnabled(Constants.IS_DEBUG_ANABLED)
       .setTag((SetupSaveGame).toString());
@@ -17,32 +16,40 @@ class SetupSaveGame{
   // -----------------------------------------------------------------------
 
   void setupFromPreviousGameState() async {
-
     liveData.targetWord.value = liveData.saveGameModel.value!.targetWord;
     liveData.wordLength.value = liveData.saveGameModel.value!.targetWord.length;
-    liveData.gameBoardStateList.value = liveData.saveGameModel.value!.gameBoardStateList;
+    liveData.gameBoardStateList.value =
+        liveData.saveGameModel.value!.gameBoardStateList;
+    liveData.gameState.value = liveData.saveGameModel.value!.gameState;
 
+    checkIfPreviousGameEnded();
   }
 
-  void save(){
+  void save() {
     if (liveData.saveGameModel.value == null) {
       var newSaveGameModel = SaveGameModel()
         ..targetWord = liveData.targetWord.value
+        ..gameState = liveData.gameState.value
         ..gameBoardStateList = liveData.gameBoardStateList;
 
       liveData.keyValueRepository.createGameData(newSaveGameModel);
     } else {
       liveData.saveGameModel.value!
         ..targetWord = liveData.targetWord.value
+        ..gameState = liveData.gameState.value
         ..gameBoardStateList = liveData.gameBoardStateList;
 
+      // liveData.keyValueRepository.updateGameData(liveData.saveGameModel.value!);
       liveData.saveGameModel.value!.save();
     }
   }
 
-  void deleteSaveGame(){
-    liveData.keyValueRepository.deleteGameData();
+  void deleteSaveGame() {
+    if (liveData.saveGameModel.value != null) {
+      liveData.keyValueRepository.deleteGameData();
+      liveData.saveGameModel.value = null;
+    }
   }
 
-
+  void checkIfPreviousGameEnded() {}
 }
